feat(app_bar): allow custom widgets in bar and simplify overflow
Applications can now provide any layout function that they want as an action
to be displayed in the app bar, though a convenience function for creating
standard icon buttons is provided.

Additionally, the overflow menu API was simplified so that it is no longer
necessary to request the close of the overflow menu when a selection is made
within it. Applications instead can check for the Tag of selected overflow
actions similarly to the current nav bar's API for indicating which nav item
was selected.