fix(nav_drawer): fix subtle multi-use animation bugs
When the nav drawer was used as both a static and a modal drawer within
the same UI (though not at the same time) its internal state would be
corrupted by calling its layout function two different ways. This fix
simply ensures that its layout function does not modify its internal
state if the drawer is not visible.