feat(bar): simplify scaled bar usage
Previously, to scale the length of the bar based on the size of
the scroll area required setting the Length property manually
as well as tedious minimum length checks and unit conversions.

This commit simplifies the API to requesting just a progress and
scale both as fractions of the available scrollable region. If
scale isn't provided a sane default is used.