This commit prefers an immediate mode API.
To that effect, the validation callback is removed and the input provides
mechanisms for the caller to manipulate error values.

Examples showcase how it is done.
One thing to note: caller needs to manually clear an error when the input
becomes valid again.

Aside: during experimentation I realised that there's two categories of state
in the input:

1. Mututally exclusive state [inactive, hovered, active, focused]
2. Optional state [errored]

Therefor "errored" is split into it's own boolean so an input can be errored in
any state. It was impossible to correctly align the input states until I realised
this.

Signed-off-by: Jack Mordaunt <jackmordaunt@gmail.com>