feat: implement simple menu and card types
This commit adds a Card implementation that can produce a surface
with a drop shadow. The card itself does not yet support rounded corners.

This Card type is then used as the foundation layer for a simple
Menu type that can present a list of widgets on a card.

This implementation is nowhere near complete, and has several problems:
- the drop shadow has artifacts at the corners where the shadow doesn't
  fade all the way to the background color.
- the drop shadow has small seams between the four different panels that
  compose it.
- the card can't itself have rounded corners without exposing an area with
  no shadow underneath.
- the MenuState currently holds the layout funcs for the menu options. I'm
  not actually certain that this is a good design, and I may want to move
  those to be provided at layout time.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>