feat: use elevation to configure card shadow
This commit switches from explicitly sizing the card shadow to
characterizing it with elevation. This more closely matches the
material design spec.

This implementation isn't great though. We still have the problem
of shadows that aren't rounded (while the thing casting them is),
and I can't yet find a means of computing the shadow from the
elevation that I'm happy with.

I think the next step is to focus on computing shadows and to
circle back to everything else once that problem is better
understood.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>