markdown: implement automatic support for bare links
If a hyperlink within a document isn't wrapped in the
markdown syntax [](), it previously wouldn't be rendered
as a link. This commit adds a lightweight heuristic that
transforms bare links into markdown hyperlinks.

I'm not certain that this is the right implementation
long-term. It may be better to pre-process the string
using separate tools, as that would save evaluating the
regular expression each frame.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>