richtext: fix text vertical offset
The previous implementation used the line height to offset
lines, and this resulted in the text's descents being
drawn beneath the bounding box returned by the dimensions.

This changes the vertical offset to only use the ascent
height to offset the current line, which brings the
offset point to the baseline of the text. The result
is that the text fits entirely within the returned
dimensions.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>