scroll: support track background color and fix bugs
Adding the track background color pointed out some alignment
bugs that required reworking the layout code significantly.

A side benefit of this work is that the click/drag hit
area of the track now extends all the way through the
padding around the track, making it significantly easier
to click the scrollbar reliably.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>