explorer: [wasm] improve large files support
Previously, reading large files (> 10MB) locks the "mainthread", making the app
unresponsive and may lead the browser to crash, due to lack of free memory.

This change fixes that behavior. Now, it loads the file using promises, which doesn't "lock the thread". This patch increases the complexity of the code, since promises can't be called from "wasm imports"/"assembly".

Using `io.Copy` will copy only 32KB of data per call, which may take some
seconds to read the entire file. However, you can still run and use the app,
since it's not block the thread.

Notice: if the file is really big (say >= 1GB) the browser may not allow to
allocate space. That is a issue between the browser and the webassembly.

Signed-off-by: Inkeliz <inkeliz@inkeliz.com>