notify: delay and synchronize default notifier creation
The default notifier instance used by the Push function was being
created with an init() function. This worked on most platforms, but
not on Android. To create a notifier on Android requires having a
JVM handle, and that isn't available until after application
startup. To handle this, I've configured the push function to
initialize the default notifier instance on first use, and to
synchronize access to it via a mutex to protect against multiple
goroutines trying to use Push simultaneously.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>