component: expand ContextArea's positioning options
This commit allows users to position a ContextArea's contextual widget
at a fixed location instead of at the location of the triggering
pointer interaction. It also allows users to provide a hint to
the ContextArea about where it is positioned within a larger
layout, enabling the ContextArea to avoid positioning
its contextual widget off-screen.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>