deps,richtext: update gio and adapt richtext to new shaper
This commit tweaks the richtext package to match the new text
shaper API in Gio core.

NOTE: this does not make the richtext package support RTL text,
but allows it to continue working for LTR languages in the new
shaper. RTL text support in this package should wait for a
more sophisticated text shaping API to be surfaced in Gio core.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>