outlay: drop previous table and rename grid to flow
This commit removes the outlay.Table type that is obsolete now
that we have a more general two dimensional grid type available.

It also renames the Grid types to use the term "Flow". This
seems like a slightly more descriptive name to me, as those types
allow a set of elements to flow along an axis until they don't
fit, at which time they are wrapped to the next line.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>