outlay: implement grid layout
This patch implements a two dimensional grid that functions much
like layout.List. The API accepts a closure to provide the dimensions
of each row and column, and a separate closure to supply the contents
of each cell.

As an implementation simplification, all rows are presumed to have
the same height as the row with index 0, though we hope to remove this
constraint in the future.

Additionally, the grid supports locking the first N rows to the top
of the grid, though it does not currently support a corresponding
feature for columns.

Co-authored-by: Chris Waldon <christopher.waldon.dev@gmail.com>
Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>