markdown: [API] revise text styling api
This commit enables consuming code to configure the text sizes
and colors generated by the markdown renderer by setting optional
fields on Renderer.Config. Additionally, the default header sizes
are now sane and no longer map 1:1 to the material design header
sizes. This commit has the side effect of dropping the dependency
on the material package.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>