component(context-area): add state update method
This commit adds a method to the context area type that permits running
the event processing independently of the layout. This is useful for applications
that need to know whether the contextual widget will be displayed during the
current frame prior to actually performing the display. If this method is
never invoked, it will automatically be invoked by Layout. If it has already
been invoked during a given frame, it will be skipped by Layout.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>