richtext: fix interactive span aliasing
This commit fixes a subtle bug that only permitted a single interactive
span per richtext.TextStyle. The loop iterating the spans copied the
spans as a loop variable and assigned an important piece of tracking
metadata only on the copy. This resulted in all interactive spans
thinking they were the first, and thus the final span was the only
active one.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>