styledtext,richtext: adapt to new text shaping APIs
This commit restructures the internals of styledtext to be compatible with
Gio's new text shaping API. Unfortunately, styledtext isn't completely bidi-
safe. Spans containing bidi text should display correctly, but if a style
change occurs within runs of text that move against the primary text
direction, the runs will be displayed out of order. The only good way to
fix this is to enable Gio core to provide a richer styled text API, which
we should be able to do pretty soon.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>