resize: maintain ratio when changing constraints
Before this change, changing the constraints (e.g. by resizing a window)
would have the following behavior:

- Increasing the max constraint would not change the absolute size of
  w1, i.e. float.Pos would be constant.

- Decreasing the max constraint would rapidly shrink the Ratio to zero.

With this change, the ratio is constant across changes in the
constraints. Increasing the constraint will lead to an absolutely larger
w1 and w2, and decreasing the constraint will lead to an absolutely
smaller w1 and w2, while maintaining a constant ratio.

Signed-off-by: Dominik Honnef <dominik@honnef.co>