styledtext: fix possible infinite loop
This commit fixes a condition in which the styledtext display could loop infinitely
if none of the text in a run could fit on a line. The text iterator would detect
zero runes of text prior to the truncator symbol and would fail to make progress,
performing the same shaping with the same runes on the next iteration. This change
ensures that if the truncator is encountered after zero runes, it is treated as
consuming a rune which must be a newline.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>