go.*,colorpicker,markdown: update to latest gio (system fonts)
This commit updates gio-x to be compatible with Gio's new text API.
In particular, the colorpicker and markdown packages now need to request
monospace fonts using the typeface instead of the variant property.

In order to allow callers to choose the monospace typeface, the colorpicker
and markdown.Renderer now expose fields allowing the developer to supply
the typeface name of the monospace font they want used.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>