component: make contextual overlay smarter
This commit updates the logic used to position the contextual widget given to
ContextArea.Layout. Previously the context area would attempt to position it
within the constraints given to the context area, but this could result in
incorrect positioning if the context area did not actually fill those
constraints within the final layout.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>