styledtext: handle too-long single words
If wrapping on words and a world longer than the max line length is encountered, we
would previously mis-handle this case and corrupt the displayed text. Now such words
are handled by allowing them to extend (invisibly) beyond the bounds of the styled
text, as we cannot intelligently break them on grapheme clusters in this package.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>