component: make menu items RTL compatible
Thanks to Imran Iqbal for prompting me to do this. Imran submitted a
great patch that manually flipped the presentation order of items based on
locale, but I really didn't want to have to do that everywhere. Instead,
I defined some layout helpers that do the right thing based on the locale
and used those to achieve the same effect for menu items. All of package
component needs this conversion, but I can't tackle the thorough testing
required right now.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>