explorer: make CreateFile API work for WASM
Earlier js.CopyBytesToJS(value, b) was used on a blank destination array. As per go docs, it copies as many bytes as min of source and destination array length, hence 0 bytes. To fix this, we just store the bytes of each Write call separately and use the Blob's feature of assembling it via parts.

Signed-off-by: Siva Dirisala <siva.dirisala@gmail.com>